+++ index.ts
@@ -319,26 +319,32 @@ async function executeSshCommands() {
                 dockerAppEnvVar += ` -e ${dockerEnvironmentVar}=${value}`;
             }
         }
-        const dockerDeploymentPort = `65505`;
-        const dockerDeploymentAppName = (dockerAppHealthCheck ? `${dockerAppName}_deploying` : dockerAppName);
-        sshCommands.push(`sudo docker run -d ${dockerAppEnvVar} --name ${dockerDeploymentAppName} -p ${dockerDeploymentPort}:${containerPort} ${dockerImageLocation}`);
+        const performDockerAppHealthCheck = dockerAppHealthCheck || !!dockerAppHealthUrls.length;
+        const actualDockerAppStartCommand = `sudo docker run -d ${dockerAppEnvVar} --name ${dockerAppName} -p ${appPublicPort}:${containerPort} ${dockerImageLocation}`;
+        if (performDockerAppHealthCheck) {
+            const dockerDeploymentPort = getRandomElement(generateWithinRange(60000, 65530, getRandomInt(1, 4)));
+            const dockerDeploymentAppName = (dockerAppHealthCheck ? `${dockerAppName}_deploying` : dockerAppName);
+            sshCommands.push(`sudo docker run -d ${dockerAppEnvVar} --name ${dockerDeploymentAppName} -p ${dockerDeploymentPort}:${containerPort} ${dockerImageLocation}`);
 
-        const localDockerAppUrl = `http://127.0.0.1:${dockerDeploymentPort}`;
-        if (!dockerAppHealthUrls.length && dockerAppHealthCheck) {
-            dockerAppHealthUrls.push(localDockerAppUrl);
+            const localDockerAppUrl = `http://127.0.0.1:${dockerDeploymentPort}`;
+            if (!dockerAppHealthUrls.length && dockerAppHealthCheck) {
+                dockerAppHealthUrls.push(localDockerAppUrl);
+            }
+            sshCommands.push(`echo Checking if app has been successfully deployed...`);
+            for (let dockerAppHealthUrl of dockerAppHealthUrls) {
+                if (dockerAppHealthUrl.startsWith("/")) dockerAppHealthUrl = localDockerAppUrl + dockerAppHealthUrl;
+                sshCommands.push(`URL="${dockerAppHealthUrl}"; __DEPLOTYN_APP_DEPLOYED__=1; for i in {1..${dockerAppHealthMaxCheck}}; do curl -sf "$URL" > /dev/null && __DEPLOTYN_APP_DEPLOYED__=0 && break || { echo "Waiting for $URL... ($i/${dockerAppHealthMaxCheck})"; sleep ${dockerAppHealthWaitTime}; }; done; echo "__DEPLOTYN_APP_DEPLOYED__=$__DEPLOTYN_APP_DEPLOYED__"`);
+            }
+            sshCommands.push(`echo App started successfully, promoting...`);
+            const fastRestartScript = `bash -lc '
+                sudo docker rm -f ${dockerAppName} && \
+                ${actualDockerAppStartCommand} && \
+                sudo docker rm -f ${dockerDeploymentAppName} 
+            '`;
+            sshCommands.push(fastRestartScript);
+        } else {   
+            sshCommands.push(actualDockerAppStartCommand);
         }
-        sshCommands.push(`echo Checking if app has been successfully deployed...`);
-        for (let dockerAppHealthUrl of dockerAppHealthUrls) {
-            if (dockerAppHealthUrl.startsWith("/")) dockerAppHealthUrl = localDockerAppUrl + dockerAppHealthUrl;
-            sshCommands.push(`URL="${dockerAppHealthUrl}"; __DEPLOTYN_APP_DEPLOYED__=1; for i in {1..${dockerAppHealthMaxCheck}}; do curl -sf "$URL" > /dev/null && __DEPLOTYN_APP_DEPLOYED__=0 && break || { echo "Waiting for $URL... ($i/${dockerAppHealthMaxCheck})"; sleep ${dockerAppHealthWaitTime}; }; done; echo "__DEPLOTYN_APP_DEPLOYED__=$__DEPLOTYN_APP_DEPLOYED__"`);
-        }
-        sshCommands.push(`echo App started successfully, promoting...`);
-        const fastRestartScript = `bash -lc '
-            sudo docker rm -f ${dockerAppName} && \
-            sudo docker run -d ${dockerAppEnvVar} --name ${dockerAppName} -p ${appPublicPort}:${containerPort} ${dockerImageLocation} && \
-            sudo docker rm -f ${dockerDeploymentAppName} 
-        '`;
-        sshCommands.push(fastRestartScript);
     }
 
     sshCommands.push("exit");
@@ -475,6 +481,35 @@ function print(action: "log" | "log!" | "error" = "log", ...content: any[]) {
         return;
     }
     console[action](...content);
+}
+
+function getRandomElement<T>(list: T[]): T {
+    if (list.length === 0) {
+        throw new Error("Cannot select from an empty array.");
+    }
+    const randomIndex = Math.floor(Math.random() * list.length);
+    return list[randomIndex];
+}
+
+function getRandomInt(min: number, max: number): number {
+    const minCeiled = Math.ceil(min);
+    const maxFloored = Math.floor(max);
+    return Math.floor(Math.random() * (maxFloored - minCeiled + 1)) + minCeiled;
+}
+
+function generateWithinRange(start: number, end: number, step: number = 1): number[] {
+    const result: number[] = [];
+
+    if (start > end) {
+        for (let i = start; i >= end; i -= Math.abs(step)) {
+            result.push(i);
+        }
+    } else {
+        for (let i = start; i <= end; i += Math.abs(step)) {
+            result.push(i);
+        }
+    }
+    return result;
 }
 
 function setupTest(argc: number, argv: string[]) {
